good rework getting close

+++ ft_make_Xx.c
@@ -6,7 +6,7 @@
 /*   By: rverscho <rverscho@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/12/20 17:41:03 by rverscho       #+#    #+#                */
-/*   Updated: 2020/01/05 19:15:36 by rverscho      ########   odam.nl         */
+/*   Updated: 2020/01/12 18:38:43 by rverscho      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void make_xx(t_flag *flags, va_list args)
 {
-	int	data;
+	long	data;
 
-	data = va_arg(args, int);
+	data = va_arg(args, long);
 	if (data < 0)
 	{
 		data = data * -1;
